+++ .vscode/launch.json
@@ -9,9 +9,9 @@
             "type": "coreclr",
             "request": "launch",
             "preLaunchTask": "build",
-            "program": "${workspaceFolder}/NotionAPI.Tests/bin/Debug/net8.0/NotionAPI.Tests.dll",
+            "program": "${workspaceFolder}/Tests/bin/Debug/net8.0/Tests.dll",
             "args": [],
-            "cwd": "${workspaceFolder}/NotionAPI.Tests",
+            "cwd": "${workspaceFolder}/Tests",
             "stopAtEntry": false,
             "console": "internalConsole"
         }
