+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
         "command": "dotnet",
         "args": [
           "build",
-          "${workspaceFolder}/NotionAPI.Tests"
+          "${workspaceFolder}/Tests"
         ],
         "problemMatcher": "$msCompile",
         "group": {
