+++ NotionAPI.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35312.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NotionAPI.Tests", "NotionAPI.Tests\NotionAPI.Tests.csproj", "{2C031E6A-3604-4540-94C6-F284778B8C30}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{2C031E6A-3604-4540-94C6-F284778B8C30}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NotionAPI", "NotionAPI\NotionAPI.csproj", "{BDD7B0CC-040A-488A-9A73-F2832DEFB2CF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NotionAPI", "NotionAPI\NotionAPI.csproj", "{BDD7B0CC-040A-488A-9A73-F2832DEFB2CF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
