+++ NotionAPI/Sources/NotionAPI.Users.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace Notion;
+namespace NotionAPI;
 
 public partial class NotionAPI
 {
