+++ Tests/Sources/Program.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using NotionAPI;
+using NotionAPI.Tests;
+
+var builder = Host.CreateApplicationBuilder(args);
+
+builder.Services
+    .AddNotionAPI("NotionAPI")
+    .AddTransient<Test>();
+
+var host = builder.Build();
+
+await host.Services.GetRequiredService<Test>().AllTestsAsync();
